requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pandas==2.1.4
numpy==1.26.2
faker==22.0.0
python-dotenv==1.0.0
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
scikit-learn==1.3.2
streamlit==1.29.0
plotly==5.18.0
python-multipart==0.0.6
